Task #3725: Change colours of the extra info icons for files/folders
Change the color of the special icons located over the files/folders in multiple views of the app, which indicate if the file/folder is saved for offline or has a public link.

The colours should be: white circle, red background and white icon image.
These should be the same colours for both available themes (dark and light).