Disable the transfer resumption for camera uploads
Temporarily disable the transfers resumption for the camera uploads service, until we have a way to receive a notification when the transfers are resumed (if there are transfers pending to resume).
This is necessary to avoid upload duplicated files, because if the camera uploads was stopped during an upload by any cause and after was activated again, in this case the transfer will be resumed but the same file will be uploaded again because it is not still in the cloud.